Replaced the standard TSE sequence with the fixed version
