+++ device_config.yaml
@@ -20,7 +20,7 @@ TxCard: !TxCard
 RxCard: !RxCard
   path: "/dev/spcm0"
   # Enable the receive channels
-  channel_enable: [1, 0, 0, 0, 0, 0, 0, 0]
+  channel_enable: [1, 1, 1, 1, 0, 0, 0, 0]
   # Set the max. amplitude per channel in mV, list length must at least match the number of enabled channels
   max_amplitude: [200, 200, 200, 200, 1000, 1000, 1000, 1000]
