+++ experiments/imaging/b0_mapping.py
@@ -16,7 +16,7 @@ from console.spcm_control.acquisition_control import AcquisitionControl
 # %%
 # Create acquisition control instance
 configuration = "../../device_config.yaml"
-acq = AcquisitionControl(configuration_file=configuration, console_log_level=logging.INFO, file_log_level=logging.DEBUG)
+acq = AcquisitionControl(configuration_file=configuration, console_log_level=logging.INFO, file_log_level=logging.DEBUG, use_star_hub=True)
 
 # %%
 # Create sequence
