+++ experiments/imaging/tse_3d.py
@@ -6,10 +6,11 @@ import logging
 import matplotlib.pyplot as plt
 import numpy as np
 
+import console.spcm_control.globals as glob
 import console.utilities.sequences as sequences
+from console.interfaces.interface_acquisition_data import AcquisitionData
+from console.interfaces.interface_acquisition_parameter import Dimensions
 from console.spcm_control.acquisition_control import AcquisitionControl
-from console.spcm_control.interface_acquisition_data import AcquisitionData
-from console.spcm_control.interface_acquisition_parameter import AcquisitionParameter, Dimensions
 
 # %%
 # Create acquisition control instance
@@ -17,92 +18,58 @@ configuration = "../../device_config.yaml"
 acq = AcquisitionControl(configuration_file=configuration, console_log_level=logging.INFO, file_log_level=logging.DEBUG)
 
 # %%
-# Construct sequence
-# dim = Dimensions(x=64, y=64, z=1)
-# dim = Dimensions(x=64, y=64, z=16)
-# dim = Dimensions(x=64, y=16, z=1)
-# dim = Dimensions(x=64, y=64, z=32)
-dim = Dimensions(x=120, y=100, z=5)
+# Create sequence
 
-ro_bw = 20e3
-te = 16e-3
-# te = 25e-3
+dim         = Dimensions(x=80, y=100, z=10)
+ro_bw       = 20e3
+te          = 20e-3
 
-seq, traj = sequences.tse.tse_3d.constructor(
-    echo_time=te,
-    repetition_time=600e-3,
-    # etl=1,
-    etl=7,
-    # gradient_correction=100e-6,
-    # gradient_correction=153e-6,
-    gradient_correction=160e-6,
-    adc_correction=0,
-    rf_duration=200e-6,
-    # fov=Dimensions(x=250e-3, y=250e-3, z=150e-3),
-    fov=Dimensions(x=240e-3, y=200e-3, z=200e-3),
-    ro_bandwidth=ro_bw,
-    n_enc=dim
+seq, traj, kdims = sequences.tse.tse_3d.constructor(
+    echo_time           = te,
+    repetition_time     = 400e-3,
+    etl                 = 10,
+    gradient_correction = 160e-6,
+    rf_duration         = 400e-6,
+    fov                 = Dimensions(x=240e-3, y=150e-3, z=150e-3),
+    channel_ro          = "y",
+    channel_pe1         = "z",
+    channel_pe2         = "x",
+    ro_bandwidth        = ro_bw,
+    n_enc               = dim
 )
 # Optional: overwrite sequence name (used to identify experiment data)
 seq.set_definition("Name", "tse_3d")
-# If z=1, image acquisition is 2D
-# seq.set_definition("Name", "tse_2d")
 
 # Calculate decimation:
-# adc_duration = dim.x / ro_bw; num_samples = adc_duration * spcm_sample_rate; decimation = num_samples / dim.x
-# => decimation = spcm_sample_rate / ro_bw
 decimation = int(acq.rx_card.sample_rate * 1e6 / ro_bw)
+glob.update_parameters(decimation = decimation)
 
 # %%
-# Larmor frequency:
-f_0 = 1965788.0
-
-# Define acquisition parameters
-params = AcquisitionParameter(
-    larmor_frequency=f_0,
-    # b1_scaling=3.4,   # 8cm sphere phantom
-    b1_scaling=3.4,
-    fov_scaling=Dimensions(
-        # Compensation of high impedance
-        x=1/0.85,
-        y=1/0.85,
-        z=1/0.85,
-        # No scaling
-        # x=1.,
-        # y=1.,
-        # z=1.,
-    ),
-    # gradient_offset=Dimensions(-200, 0, 0),
-    decimation=decimation,
-
-    # num_averages=10,
-    # averaging_delay=1,
-)
+#Unroll and run sequence
 
 # Perform acquisition
-acq.set_sequence(parameter=params, sequence=seq)
-
-
-
-# %%
+acq.set_sequence(sequence=seq)
 acq_data: AcquisitionData = acq.run()
 
-ksp = sequences.tse_3d.sort_kspace(acq_data.raw, trajectory=traj, dim=dim)
-ksp = ksp.squeeze()
-
-
-# fig, ax = plt.subplots(1, 1)
-# _ = ax.plot(np.abs(ksp).T)
-# np.argmax(np.abs(ksp), axis = 1)
+# %%
+#sort data in to kspace array
+ksp = sequences.tse.tse_3d.sort_kspace(acq_data.raw, traj, kdims).squeeze()
 
 # %%
-# img = np.fft.fftshift(np.fft.fftn(np.fft.fftshift(ksp)))
-img = np.fft.ifftshift(np.fft.fftn(ksp))
+#plot data
+if len(np.shape(ksp)) == 4:
+    img = np.fft.fftshift(np.fft.fftn(np.fft.fftshift(ksp[0,...])))
+else:
+    img = np.fft.fftshift(np.fft.fftn(np.fft.fftshift(ksp)))
+
 
 idx = int(img.shape[0]/2)
 fig, ax = plt.subplots(1, 2, figsize=(8, 4))
-if len(img.shape) == 3:
+if len(ksp.shape) == 3:
     ax[0].imshow(np.abs(ksp[idx, ...]), cmap="gray")
+    ax[1].imshow(np.abs(img[idx, ...]), cmap="gray")
+elif len(ksp.shape) == 4:
+    ax[0].imshow(np.abs(ksp[0,idx, ...]), cmap="gray")
     ax[1].imshow(np.abs(img[idx, ...]), cmap="gray")
 else:
     ax[0].imshow(np.abs(ksp), cmap="gray")
@@ -124,7 +91,7 @@ for k, x in enumerate(img[:, ...]):
     ax[k].imshow(np.abs(x), vmin=total_min, vmax=total_max, cmap="gray")
     ax[k].axis("off")
 _ = [a.remove() for a in ax[k+1:]]
-fig.set_tight_layout(tight=0.)
+fig.set_tight_layout(True)
 fig.set_facecolor("black")
 
 
@@ -132,22 +99,19 @@ fig.set_facecolor("black")
 # %%
 
 acq_data.add_info({
-    "subject": "brain_slice, tse_david, miteq_preamp - noise",
+    "subject": "Birdcage - In-vivo",
     "echo_time": te,
     "dim": [dim.x, dim.y, dim.z],
-    # "subject": "brain-slice",
     # "sequence_info": "etl = 7, optimized grad correction",
 })
 
 acq_data.add_data({
-    "trajectory": traj,
+    "trajectory": np.array(traj),
     "kspace": ksp,
     "image": img
 })
 
-acq_data.save(save_unprocessed=True, user_path=r"C:\Users\Tom\Desktop\spcm-data\20240227 - SNR tests")
-#acq_data.save(save_unprocessed=True, user_path=r"C:\Users\Tom\Desktop\spcm-data\in-vivo")
-# acq_data.save(save_unprocessed=True, user_path=r"C:\Users\Tom\Desktop\spcm-data\b0-map")
-# acq_data.save(save_unprocessed=True, user_path=r"C:\Users\Tom\Desktop\spcm-data\brain-slice")
+acq_data.save(save_unprocessed=False, user_path=r"C:\Users\Tom\Desktop\spcm-data\20240320 - Birdcage")
+
 # %%
 del acq
